config/composes/new_client_app/requirements.txt
textual==0.53.1
python-dotenv
mysql-connector-python
